Consolidate share consumer constraint tests into integration tests
  - Replace Thread.sleep() with Awaitility pattern that verifies poll blocking via trace log
  - Add Mockito spy on LogAccessor to confirm the logging message
  - Migrate shouldHandlePartialAcknowledgmentCorrectly test from ConstraintTests
  - Migrate shouldHandleConcurrentAcknowledgmentAttempts test from ConstraintTests
  - Fix topic collision by assigning unique topics to each test
  - Remove duplicate ShareKafkaMessageListenerContainerConstraintTests class

  This improves test reliability by verifying actual blocking behavior rather than
  assuming Thread.sleep duration is sufficient, and eliminates test duplication.
